project structure refactor based on clean architecture
 added lint package and some rules

 changed repositories and interfaces names to follow conventions

added empty files to complete structure

 corrected layer dependency
