+++ analysis_options.yaml
@@ -22,6 +22,12 @@ linter:
   # `// ignore_for_file: name_of_lint` syntax on the line or in the file
   # producing the lint.
   rules:
+    - annotate_overrides
+    - await_only_futures
+    - non_constant_identifier_names
+    - prefer_relative_imports
+    - avoid_empty_else
+    - empty_constructor_bodies
     # avoid_print: false  # Uncomment to disable the `avoid_print` rule
     # prefer_single_quotes: true  # Uncomment to enable the `prefer_single_quotes` rule
