+++ lib/data/repository/genre_repository.dart
@@ -1,16 +1,17 @@
 import 'dart:convert';
 
 import 'package:flutter/services.dart';
-import '../entities/genre.dart';
+import '../../domain/repository/genre_repository_interface.dart';
+import '../../domain/entity/genre.dart';
 
-class GenreRepository {
+class GenreRepositoryImp implements IGenreRepository {
+  @override
   Future<List<Genre>> getGenreNamesById(List<num> ids) async {
     final String response =
         await rootBundle.loadString('assets/mock_data/genres.json');
     List<dynamic> genres = json.decode(response)['genres'];
     return genres
         .where((genre) => ids.contains(genre['id']))
-        .toList()
         .map((gen) => Genre.fromJson(gen))
         .toList();
   }
