+++ lib/domain/repository/genre_repository_interface.dart
@@ -1,0 +1,5 @@
+import '../entity/genre.dart';
+
+abstract class IGenreRepository{
+  Future<List<Genre>> getGenreNamesById(List<num> ids);
+}
