+++ lib/domain/repository/movie_repository_interface.dart
@@ -1,0 +1,5 @@
+import '../entity/movie.dart';
+
+abstract class IMovieRepository {
+  Future<List<Movie>> getMovies();
+}
