+++ lib/presentation/widget/movie_backdrop_and_title_stack.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'title_widget.dart';
 
-import '../ui_constants.dart';
+import '../../core/util/constants/ui_constants.dart';
 
 class MovieBackdropAndTitleStack extends StatelessWidget {
   const MovieBackdropAndTitleStack({
