+++ lib/presentation/widget/poster_and_overview_stack.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../ui_constants.dart';
+import '../../core/util/constants/ui_constants.dart';
 import 'fav_button_counter_column.dart';
 import 'movie_overview.dart';
