+++ lib/presentation/widget/release_date_and_vote_average_row.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../ui_constants.dart';
+import '../../core/util/constants/ui_constants.dart';
 import 'middle_text.dart';
 
 class ReleaseDateAndVoteAverageRow extends StatelessWidget {
