+++ lib/presentation/widget/row_genre_item.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-import '../text_styles/custom_text_style.dart';
-import '../ui_constants.dart';
+import '../../core/util/text_styles/custom_text_style.dart';
+import '../../core/util/constants/ui_constants.dart';
 
 class RowGenreItem extends StatelessWidget {
   const RowGenreItem({super.key, required this.genre});
