+++ lib/presentation/widget/title_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../text_styles/custom_text_style.dart';
+import '../../core/util/text_styles/custom_text_style.dart';
 
 class TitleWidget extends StatelessWidget {
   const TitleWidget({
